python/requirements.txt
azure-cosmos==4.7.0
openai
